RTCM Raise Error if more data is needed than is in the payload
A partial message will not be returned, previously all the remaining attributes would have been returned as 0.

This change makes it that some tests fail, (GPS MSM processing) since they were passing due to the side effect of fields without data in the payload being returned as 0's. .

Also fixed the rtcmfile.py, it said enter 3 for raise and stop, ERR_RAISE is 2 not 3.